% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{campus_market}
\alias{campus_market}
\title{Campus market}
\format{A data frame with ...}
\usage{
data(campus_market)
}
\description{
Campus market
}
\details{
Description provided in attr(campus_market,"description")
}
\keyword{datasets}

