% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{cf}
\alias{cf}
\title{Movie ratings}
\format{A data frame with 110 rows and 4 variables}
\usage{
data(cf)
}
\description{
Movie ratings
}
\details{
Use collaborative filtering to create recommendations based on ratings from existing users. Description provided in attr(df,"description")
}
\keyword{datasets}
