% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{confint_robust}
\alias{confint_robust}
\title{Confidence interval for robust estimators}
\usage{
confint_robust(object, level = 0.95, dist = "norm", vcov = NULL, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{level}{The confidence level required}

\item{dist}{Distribution to use ("norm" or "t")}

\item{vcov}{Covariance matrix generated by, e.g., sandwich::vcovHC}

\item{...}{Additional argument(s) for methods}
}
\description{
Confidence interval for robust estimators
}
\details{
Wrapper for confint with robust standard errors. See \url{https://stackoverflow.com/questions/3817182/vcovhc-and-confidence-interval/3820125#3820125}
}
