% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{crs}
\alias{crs}
\title{Collaborative Filtering}
\usage{
crs(dataset, id, prod, pred, rate, data_filter = "")
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{id}{String with name of the variable containing user ids}

\item{prod}{String with name of the variable with product ids}

\item{pred}{Products to predict for}

\item{rate}{String with name of the variable with product ratings}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "training == 1")}
}
\value{
A data.frame with the original data and a new column with predicted ratings
}
\description{
Collaborative Filtering
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/crs.html} for an example in Radiant
}
