% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtree.R
\name{cv.crtree}
\alias{cv.crtree}
\title{Cross-validation for Classification and Regression Trees}
\usage{
cv.crtree(object, K = 5, repeats = 1, cp, pcp = seq(0, 0.01,
  length.out = 11), seed = 1234, trace = TRUE, fun, ...)
}
\arguments{
\item{object}{Object of type "rpart" or "crtree" to use as a starting point for cross validation}

\item{K}{Number of cross validation passes to use}

\item{repeats}{Number of times to repeat the K cross-validation steps}

\item{cp}{Complexity parameter used when building the (e.g., 0.0001)}

\item{pcp}{Complexity parameter to use for pruning}

\item{seed}{Random seed to use as the starting point}

\item{trace}{Print progress}

\item{fun}{Function to use for model evaluation (e.g., auc for classification or RMSE for regression)}

\item{...}{Additional arguments to be passed to 'fun'}
}
\value{
A data.frame sorted by the mean, sd, min, and max of the performance metric
}
\description{
Cross-validation for Classification and Regression Trees
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/crtreenn.html} for an example in Radiant
}
\seealso{
\code{\link{crtree}} to generate an initial model that can be passed to cv.crtree
}
