% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{direct_marketing}
\alias{direct_marketing}
\title{Direct marketing}
\format{A data frame with ...}
\usage{
data(direct_marketing)
}
\description{
Direct marketing
}
\details{
Description provided in attr(direct_marketing,"description")
}
\keyword{datasets}

