% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{.as_int}
\alias{.as_int}
\title{Convenience function used in "simulater"}
\usage{
.as_int(x, dataset = list())
}
\arguments{
\item{x}{String to be converted to an integer}

\item{dataset}{Data list}
}
\value{
An integer
}
\description{
Convenience function used in "simulater"
}
