% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{.as_num}
\alias{.as_num}
\title{Convenience function used in "simulater"}
\usage{
.as_num(x, dataset = list())
}
\arguments{
\item{x}{String to be converted to an numeric value}

\item{dataset}{Data list}
}
\value{
An numeric value
}
\description{
Convenience function used in "simulater"
}
