% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalreg.R
\name{evalreg}
\alias{evalreg}
\title{Model evalreg}
\usage{
evalreg(dataset, pred, rvar, train = "", data_filter = "")
}
\arguments{
\item{dataset}{Dataset name (string). This can be a dataframe in the global environment or an element in an r_data list from Radiant}

\item{pred}{Predictions or predictors}

\item{rvar}{Response variable}

\item{train}{Use data from training ("Training"), validation ("Validation"), both ("Both"), or all data ("All") to evaluate model evalreg}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}
}
\value{
A list of results
}
\description{
Model evalreg
}
\details{
See \url{http://radiant-rstats.github.io/docs/model/evalreg.html} for an example in Radiant
}
\seealso{
\code{\link{summary.evalreg}} to summarize results

\code{\link{plot.evalreg}} to plot results
}

