% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{heights}
\alias{heights}
\title{Heights}
\format{A data frame with ...}
\usage{
data(heights)
}
\description{
Heights
}
\details{
Description provided in attr(heights,"description")
}
\keyword{datasets}

