% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{houseprices}
\alias{houseprices}
\title{Houseprices}
\format{A data frame with ...}
\usage{
data(houseprices)
}
\description{
Houseprices
}
\details{
Description provided in attr(houseprices,"description")
}
\keyword{datasets}

