% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{ideal}
\alias{ideal}
\title{Ideal data for linear regression}
\format{A data frame with 1,000 rows and 4 variables}
\usage{
data(ideal)
}
\description{
Ideal data for linear regression
}
\details{
Description provided in attr(ideal,  "description")
}
\keyword{datasets}
