% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{ketchup}
\alias{ketchup}
\title{Data on ketchup choices}
\format{
A data frame with 2,798 rows and 14 variables
}
\usage{
data(ketchup)
}
\description{
Data on ketchup choices
}
\details{
Choice behavior for a sample of 300 individuals in a panel of households in Springfield, Missouri (USA). Description provided in attr(ketchup,"description")
}
\keyword{datasets}
