% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalreg.R
\name{MAE}
\alias{MAE}
\title{Mean Absolute Error}
\usage{
MAE(pred, rvar)
}
\arguments{
\item{pred}{Prediction (vector)}

\item{rvar}{Response (vector)}
}
\value{
Mean Absolute Error
}
\description{
Mean Absolute Error
}
