% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{minmax}
\alias{minmax}
\title{Calculate min and max before standardization}
\usage{
minmax(dataset)
}
\arguments{
\item{dataset}{Data frame}
}
\value{
Data frame min and max attributes
}
\description{
Calculate min and max before standardization
}
