% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{movie_contract}
\alias{movie_contract}
\title{Movie contract decision tree}
\format{A nested list for decision and chance nodes, probabilities and payoffs}
\usage{
data(movie_contract)
}
\description{
Movie contract decision tree
}
\details{
Use decision analysis to create a decision tree for an actor facing a contract decision
}
\keyword{datasets}
