% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{onehot}
\alias{onehot}
\title{One hot encoding of data.frames}
\usage{
onehot(dataset, all = FALSE, df = FALSE)
}
\arguments{
\item{dataset}{Dataset to endcode}

\item{all}{Extract all factor levels (e.g., for tree-based models)}

\item{df}{Return a data.frame (tibble)}
}
\description{
One hot encoding of data.frames
}
\examples{
head(onehot(diamonds, df = TRUE))
head(onehot(diamonds, all = TRUE, df = TRUE))
}
