% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{pdp_plot}
\alias{pdp_plot}
\title{Create Partial Dependence Plots}
\usage{
pdp_plot(
  x,
  plot_list = list(),
  incl,
  incl_int,
  fix = TRUE,
  hline = TRUE,
  nr = 20,
  minq = 0.025,
  maxq = 0.975
)
}
\arguments{
\item{x}{Return value from a model}

\item{plot_list}{List used to store plots}

\item{incl}{Which variables to include in PDP plots}

\item{incl_int}{Which interactions to investigate in PDP plots}

\item{fix}{Set the desired limited on yhat or have it calculated automatically.
Set to FALSE to have y-axis limits set by ggplot2 for each plot}

\item{hline}{Add a horizontal line at the average of the target variable. When set to FALSE
no line is added. When set to a specific number, the horizontal line will be added at that value}

\item{nr}{Number of values to use to generate predictions for a numeric explanatory variable}

\item{minq}{Quantile to use for the minimum value for simulation of numeric variables}

\item{maxq}{Quantile to use for the maximum value for simulation of numeric variables}
}
\description{
Create Partial Dependence Plots
}
