% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalbin.R
\name{plot.confusion}
\alias{plot.confusion}
\title{Plot method for the confusion matrix}
\usage{
\method{plot}{confusion}(x, scale_y = FALSE, shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{evalreg}}}

\item{scale_y}{Free scale in faceted plot of the confusion matrix (TRUE or FALSE)}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the confusion matrix
}
\details{
See \url{http://radiant-rstats.github.io/docs/model/evalreg.html} for an example in Radiant
}
\seealso{
\code{\link{evalreg}} to generate results

\code{\link{summary.evalreg}} to summarize results
}

