% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{plot.crs}
\alias{plot.crs}
\title{Plot method for the crs function}
\usage{
\method{plot}{crs}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{crs}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the crs function
}
\details{
Plot that compares actual to predicted ratings. See \url{https://radiant-rstats.github.io/docs/model/crs.html} for an example in Radiant
}
\seealso{
\code{\link{crs}} to generate results

\code{\link{summary.crs}} to summarize results
}
