% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalreg.R
\name{plot.evalreg}
\alias{plot.evalreg}
\title{Plot method for the evalreg function}
\usage{
\method{plot}{evalreg}(x, shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{evalreg}}}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the evalreg function
}
\details{
See \url{http://vnijs.github.io/radiant/analytics/evalreg.html} for an example in Radiant
}
\seealso{
\code{\link{evalreg}} to generate results

\code{\link{summary.evalreg}} to summarize results
}

