% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{plot.logistic}
\alias{plot.logistic}
\title{Plot method for the logistic function}
\usage{
\method{plot}{logistic}(x, plots = "", conf_lev = 0.95, intercept = FALSE,
  shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{logistic}}}

\item{plots}{Plots to produce for the specified GLM model. Use "" to avoid showing any plots (default). "hist" shows histograms of all variables in the model. "scatter" shows scatter plots (or box plots for factors) for the response variable with each explanatory variable. "dashboard" is a series of four plots used to visually evaluate model. "coef" provides a coefficient plot}

\item{conf_lev}{Confidence level to use for coefficient and odds confidence intervals (.95 is the default)}

\item{intercept}{Include the intercept in the coefficient plot (TRUE or FALSE). FALSE is the default}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the logistic function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}
\examples{
result <- logistic("titanic", "survived", c("pclass","sex"), lev = "Yes")
plot(result, plots = "coef")

}
\seealso{
\code{\link{logistic}} to generate results

\code{\link{plot.logistic}} to plot results

\code{\link{predict.logistic}} to generate predictions

\code{\link{plot.logit_predict}} to plot prediction output
}

