% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{predict.logistic}
\alias{predict.logistic}
\title{Predict method for the logistic function}
\usage{
\method{predict}{logistic}(object, pred_vars = "", pred_data = "",
  pred_cmd = "", prn = 100, se = FALSE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{logistic}}}

\item{pred_vars}{Variables selected to generate predictions}

\item{pred_data}{Provide the name of a dataframe to generate predictions (e.g., "titanic"). The dataset must contain all columns used in the estimation}

\item{pred_cmd}{Generate predictions using a command. For example, `pclass = levels(pclass)` would produce predictions for the different levels of factor `pclass`. To add another variable use a `,` (e.g., `pclass = levels(pclass), age = seq(0,100,20)`)}

\item{prn}{Number of lines of prediction results to print. Nothing is printed if prn is 0. Use -1 to print all lines (default).}

\item{se}{Logical that indicates if prediction standard errors should be calculated (default = FALSE)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Predict method for the logistic function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}
\examples{
result <- logistic("titanic", "survived", c("pclass","sex"), lev = "Yes")
 predict(result, pred_cmd = "pclass = levels(pclass)")
logistic("titanic", "survived", c("pclass","sex"), lev = "Yes") \%>\%
  predict(pred_cmd = "sex = c('male','female')")
logistic("titanic", "survived", c("pclass","sex"), lev = "Yes") \%>\%
 predict(pred_data = "titanic")

}
\seealso{
\code{\link{logistic}} to generate the result

\code{\link{summary.logistic}} to summarize results

\code{\link{plot.logistic}} to plot results

\code{\link{plot.logit_predict}} to plot prediction output
}

