% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{predict.naiveBayes}
\alias{predict.naiveBayes}
\title{Place holder for predict.naiveBayes until e1071 is updated to allow logical in prediction}
\usage{
\method{predict}{naiveBayes}(object, newdata, type = c("class", "raw"),
  threshold = 0.001, eps = 0, ...)
}
\arguments{
\item{object}{An object of class \code{"naiveBayes"}.}

\item{newdata}{A dataframe with new predictors (with possibly fewer
columns than the training data). Note that the column names of
\code{newdata} are matched against the training data ones.}

\item{type}{If \code{"raw"}, the conditional a-posterior
probabilities for each class are returned, and the class with
maximal probability else}

\item{threshold}{Value replacing cells with probabilities within \code{eps} range.}

\item{eps}{Double for specifying an epsilon-range to apply laplace
smoothing (to replace zero or close-zero probabilities by \code{theshold}.)}

\item{...}{Additional arguments}
}
\description{
Place holder for predict.naiveBayes until e1071 is updated to allow logical in prediction
}

