% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{print.model.predict}
\alias{print.model.predict}
\title{Print method for the model prediction}
\usage{
\method{print}{model.predict}(x, ..., n = 10, header = "", lev = "")
}
\arguments{
\item{x}{Return value from prediction method}

\item{...}{further arguments passed to or from other methods}

\item{n}{Number of lines of prediction results to print. Use -1 to print all lines}

\item{header}{Header line}

\item{lev}{The level in the response variable defined as _success_ for classification models}
}
\description{
Print method for the model prediction
}

