% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest.R
\name{print.rforest.predict}
\alias{print.rforest.predict}
\title{Print method for predict.rforest}
\usage{
\method{print}{rforest.predict}(x, ..., n = 10)
}
\arguments{
\item{x}{Return value from prediction method}

\item{...}{further arguments passed to or from other methods}

\item{n}{Number of lines of prediction results to print. Use -1 to print all lines}
}
\description{
Print method for predict.rforest
}
