% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{print_predict_model}
\alias{print_predict_model}
\title{Print method for the model prediction}
\usage{
print_predict_model(x, ..., n = 10, header = "")
}
\arguments{
\item{x}{Return value from prediction method}

\item{...}{further arguments passed to or from other methods}

\item{n}{Number of lines of prediction results to print. Use -1 to print all lines}

\item{header}{Header line}
}
\description{
Print method for the model prediction
}
