% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{radiant.model-deprecated}
\alias{glm_reg}
\alias{performance}
\alias{radiant.model-deprecated}
\alias{radiant.model-deprecated-package}
\alias{regression}
\title{Deprecated function(s) in the radiant.model package}
\usage{
regression(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
radiant. They will eventually be  removed.
}
\section{Details}{

\tabular{rl}{
  \code{regression} is now a synonym for \code{\link{regress}}\cr
  \code{glm_reg} is now a synonym for \code{\link{logistic}}\cr
  \code{performance} is now a synonym for \code{\link{evalbin}}\cr
}
}

