% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.model}
\alias{radiant.model}
\alias{radiant.model-package}
\title{radiant.model}
\usage{
radiant.model()
}
\description{
radiant.model

Launch Radiant in the default browser
}
\details{
See \url{http://vnijs.github.io/radiant} for documentation and tutorials
}

