% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{remove_comments}
\alias{remove_comments}
\title{Remove comments from formula before it is evaluated}
\usage{
remove_comments(x)
}
\arguments{
\item{x}{Input string}
}
\value{
Cleaned string
}
\description{
Remove comments from formula before it is evaluated
}
