% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render.DiagrammeR}
\alias{render.DiagrammeR}
\title{Method to render DiagrammeR plots}
\usage{
\method{render}{DiagrammeR}(object, ...)
}
\arguments{
\item{object}{DiagrammeR plot}

\item{...}{Additional arguments}
}
\description{
Method to render DiagrammeR plots
}
