% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{repeater}
\alias{repeater}
\title{Repeated simulation}
\usage{
repeater(nr = 12, vars = "", grid = "", sum_vars = "", byvar = "sim",
  fun = "sum_rm", form = "", seed = NULL, name = "", sim = "")
}
\arguments{
\item{nr}{Number times to repeat the simulation}

\item{vars}{Variables to use in repeated simulation}

\item{grid}{Character vector of expressions to use in grid search for constants}

\item{sum_vars}{(Numeric) variables to summaries}

\item{byvar}{Variable(s) to group data by before summarizing}

\item{fun}{Functions to use for summarizing}

\item{form}{A character vector with the formula to apply to the summarized data}

\item{seed}{Seed for the repeated simulation}

\item{name}{Name for data.frame with the repeated simulation data (optional)}

\item{sim}{Return value from the simulater function (data.frame or data.frame name)}
}
\description{
Repeated simulation
}
\examples{
result <- simulater(
  const = c("var_cost 5","fixed_cost 1000"), 
  norm = "E 0 100;",
  discrete = "price 6 8 .3 .7;",
  form = c(
    "demand = 1000 - 50*price + E", 
    "profit = demand*(price-var_cost) - fixed_cost",
    "profit_small = profit < 100"
  ), 
  seed = 1234
)

repeater(
  nr = 12, 
  vars = c("E","price"), 
  sum_vars = "profit", 
  byvar = "sim", 
  form = "profit_365 = profit < 36500", 
  seed = 1234, 
  sim = result
) \%>\% head

}
