% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{scaledf}
\alias{scaledf}
\title{Center or standardize variables in a data frame}
\usage{
scaledf(dat, center = TRUE, scale = TRUE, sf = 2, wts = NULL,
  calc = TRUE)
}
\arguments{
\item{dat}{Data frame}

\item{center}{Center data (TRUE or FALSE)}

\item{scale}{Scale data (TRUE or FALSE)}

\item{sf}{Scaling factor (default is 2)}

\item{wts}{Weights to use (default is NULL for no weights)}

\item{calc}{Calculate mean and sd or use attributes attached to dat}
}
\value{
Scaled data frame
}
\description{
Center or standardize variables in a data frame
}
\seealso{
\code{\link{copy_attr}} to copy attributes from a traning to a validation dataset
}
