% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{sensitivity.dtree}
\alias{sensitivity.dtree}
\title{Evaluate sensitivity of the decision tree}
\usage{
\method{sensitivity}{dtree}(object, vars = NULL, decs = NULL,
  shiny = FALSE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{dtree}}}

\item{vars}{Variables to include in the sensitivity analysis}

\item{decs}{Decisions to include in the sensitivity analysis}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{Additional arguments}
}
\description{
Evaluate sensitivity of the decision tree
}
\details{
See \url{http://radiant-rstats.github.io/docs/model/dtree.html} for an example in Radiant
}

