% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{sensitivity}
\alias{sensitivity}
\title{Method to evaluate sensitivity of an analysis}
\usage{
sensitivity(object, ...)
}
\arguments{
\item{object}{Object of relevant class for which to evaluate sensitivity}

\item{...}{Additional arguments}
}
\description{
Method to evaluate sensitivity of an analysis
}

