% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{sim_summary}
\alias{sim_summary}
\title{Print simulation summary}
\usage{
sim_summary(dataset, dc = get_class(dataset), fun = "", dec = 4)
}
\arguments{
\item{dataset}{Simulated data}

\item{dc}{Variable classes}

\item{fun}{Summary function to apply}

\item{dec}{Number of decimals to show}
}
\description{
Print simulation summary
}
