% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{store.crs}
\alias{store.crs}
\title{Store predicted values generated in the crs function}
\usage{
\method{store}{crs}(object, name = "predict_cf", envir = parent.frame(),
  ...)
}
\arguments{
\item{object}{Return value from crs}

\item{name}{Name of the dataset to store}

\item{envir}{Environment to assign 'new' dataset (optional). Used when an r_data list is not available}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store predicted values generated in the crs function
}
\details{
Store data frame with predictions in Radiant r_data list if available. See \url{https://radiant-rstats.github.io/docs/model/crs.html} for an example in Radiant
}
