% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{store.logistic.predict}
\alias{store.logistic.predict}
\title{Store predicted values generated in the logistic function}
\usage{
\method{store}{logistic.predict}(object, ..., name = "predict_logit")
}
\arguments{
\item{object}{Return value from \code{\link{logistic}} or \code{\link{predict.logistic}}}

\item{...}{Additional arguments. Must include data or dataset name (e.g., data = mtcars or data = "mtcars")}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Store predicted values generated in the logistic function
}
\details{
Use \code{\link{store.logistic.predict}} or \code{\link{store.logistic}} instead

See \url{http://vnijs.github.io/radiant/quant/glm_reg.html} for an example in Radiant
}

