% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{store.logistic}
\alias{store.logistic}
\title{Store residuals from a model generated in the logistic function}
\usage{
\method{store}{logistic}(object, ..., name = "residuals_logit")
}
\arguments{
\item{object}{Return value from \code{\link{logistic}}}

\item{...}{Additional arguments}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store residuals from a model generated in the logistic function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/logistic} for an example in Radiant
}

