% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl.R
\name{store.mnl.predict}
\alias{store.mnl.predict}
\title{Store predicted values generated in the mnl function}
\usage{
\method{store}{mnl.predict}(dataset, object, name = "pred_mnl", ...)
}
\arguments{
\item{dataset}{Dataset to add predictions to}

\item{object}{Return value from model function}

\item{name}{Variable name(s) assigned to predicted values. If empty, the levels of the response variable will be used}

\item{...}{Additional arguments}
}
\description{
Store predicted values generated in the mnl function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/mnl.html} for an example in Radiant
}
\examples{
result <- mnl(titanic, rvar = "survived", evar = c("pclass", "sex", "age"))
pred <- predict(result, pred_data = titanic)
titanic <- store(titanic, pred, name = c("Yes", "No"))

}
