% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{store.model}
\alias{store.model}
\title{Store residuals from a model}
\usage{
\method{store}{model}(object, ..., name = "residuals")
}
\arguments{
\item{object}{Return value from a model function}

\item{...}{Additional arguments}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store residuals from a model
}
\details{
See \url{http://radiant-rstats.github.io/docs/model/regress.html} for an example in Radiant
}
\examples{
regress(diamonds, rvar = "price", evar = c("carat","cut")) \%>\%
  store \%>\% head

}

