% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{store.regress.predict}
\alias{store.regress.predict}
\title{Store predicted values generated in the regress function}
\usage{
\method{store}{regress.predict}(object, ..., name = "pred_reg")
}
\arguments{
\item{object}{Return value from \code{\link{predict.regress}}}

\item{...}{Additional arguments. Must include data or dataset name (e.g., data = mtcars or data = "mtcars")}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store predicted values generated in the regress function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}

