% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{store.regress}
\alias{store.regress}
\title{Store residuals from a model generated in the regress function}
\usage{
\method{store}{regress}(object, ..., name = "residuals_reg")
}
\arguments{
\item{object}{Return value from \code{\link{regress}}}

\item{...}{Additional arguments}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store residuals from a model generated in the regress function
}
\details{
See \url{http://vnijs.github.io/radiant/quant/regression.html} for an example in Radiant
}

