% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{store_ann}
\alias{store_ann}
\title{Deprecated function to store predictions from an ANN}
\usage{
store_ann(object, data = object$dataset, name = paste0("predict_ann"))
}
\arguments{
\item{object}{Return value from \code{\link{predict.ann}}}

\item{data}{Dataset name}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Deprecated function to store predictions from an ANN
}
\details{
Use \code{\link{store.model.predict}} or \code{\link{store.model}} instead
}
