% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{store_glm}
\alias{store_glm}
\title{Deprecated function to store logistic regression residuals and predictions}
\usage{
store_glm(object, data = object$dataset, type = "residuals",
  name = paste0(type, "_logit"))
}
\arguments{
\item{object}{Return value from \code{\link{logistic}} or \code{\link{predict.logistic}}}

\item{data}{Dataset name}

\item{type}{Residuals ("residuals") or predictions ("predictions"). For predictions the dataset name must be provided}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Deprecated function to store logistic regression residuals and predictions
}
\details{
Use \code{\link{store.model.predict}} or \code{\link{store.model}} instead
}

