% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{store_reg}
\alias{store_reg}
\title{Deprecated function to store regression residuals and predictions}
\usage{
store_reg(object, data = object$dataset, type = "residuals",
  name = paste0(type, "_reg"))
}
\arguments{
\item{object}{Return value from \code{\link{regress}} or \code{\link{predict.regress}}}

\item{data}{Dataset name}

\item{type}{Residuals ("residuals") or predictions ("predictions"). For predictions the dataset name must be provided}

\item{name}{Variable name assigned to the residuals or predicted values}
}
\description{
Deprecated function to store regression residuals and predictions
}
\details{
Use \code{\link{store.model.predict}} or \code{\link{store.model}} instead
}
