% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{summary.avar}
\alias{summary.avar}
\title{Summary method for the avar function}
\usage{
\method{summary}{avar}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{avar}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the avar function
}
\details{
One-Way Anova function ...
}
