% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{summary.crs}
\alias{summary.crs}
\title{Summary method for Collaborative Filter}
\usage{
\method{summary}{crs}(object, n = 36, dec = 2, ...)
}
\arguments{
\item{object}{Return value from \code{\link{crs}}}

\item{n}{Number of lines of recommendations to print. Use -1 to print all lines}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for Collaborative Filter
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/crs.html} for an example in Radiant
}
\examples{
crs(ratings, id = "Users", prod = "Movies", pred = c("M6", "M7", "M8", "M9", "M10"),
    rate = "Ratings", data_filter = "training == 1") \%>\% summary()

}
\seealso{
\code{\link{crs}} to generate the results

\code{\link{plot.crs}} to plot results if the actual ratings are available
}
