% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{summary.dtree}
\alias{summary.dtree}
\title{Summary method for the dtree function}
\usage{
\method{summary}{dtree}(object, input = TRUE, output = FALSE,
  dec = 2, ...)
}
\arguments{
\item{object}{Return value from \code{\link{simulater}}}

\item{input}{Print decision tree input}

\item{output}{Print decision tree output}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the dtree function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/dtree.html} for an example in Radiant
}
\examples{
dtree(movie_contract, opt = "max") \%>\% summary(input = TRUE)
dtree(movie_contract, opt = "max") \%>\% summary(input = FALSE, output = TRUE)

}
\seealso{
\code{\link{dtree}} to generate the results

\code{\link{plot.dtree}} to plot results

\code{\link{sensitivity.dtree}} to plot results
}
