% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalbin.R
\name{summary.evalbin}
\alias{summary.evalbin}
\title{Summary method for the evalbin function}
\usage{
\method{summary}{evalbin}(object, prn = TRUE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{evalbin}}}

\item{prn}{Print full table of measures per model and bin}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the evalbin function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/evalbin.html} for an example in Radiant
}
\examples{
evalbin("titanic", "age", "survived") \%>\% summary
evalbin("titanic", c("age","fare"), "survived") \%>\% summary

}
\seealso{
\code{\link{evalbin}} to summarize results

\code{\link{plot.evalbin}} to plot results
}
