% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalreg.R
\name{summary.evalreg}
\alias{summary.evalreg}
\title{Summary method for the evalreg function}
\usage{
\method{summary}{evalreg}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{evalreg}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the evalreg function
}
\details{
See \url{http://radiant-rstats.github.io/docs/model/evalreg.html} for an example in Radiant
}
\seealso{
\code{\link{evalreg}} to summarize results

\code{\link{plot.evalreg}} to plot results
}

