% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{summary.nb}
\alias{summary.nb}
\title{Summary method for the nb function}
\usage{
\method{summary}{nb}(object, dec = 3, ...)
}
\arguments{
\item{object}{Return value from \code{\link{nb}}}

\item{dec}{Decimals}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the nb function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/nb.html} for an example in Radiant
}
\examples{
result <- nb("titanic", "survived", c("pclass","sex","age"))
summary(result)

}
\seealso{
\code{\link{nb}} to generate results

\code{\link{plot.nb}} to plot results

\code{\link{predict.nb}} for prediction
}
