% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{summary.repeater}
\alias{summary.repeater}
\title{Summarize repeated simulation}
\usage{
\method{summary}{repeater}(object, dec = 4, ...)
}
\arguments{
\item{object}{Return value from \code{\link{repeater}}}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summarize repeated simulation
}

