% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest.R
\name{summary.rforest}
\alias{summary.rforest}
\title{Summary method for the rforest function}
\usage{
\method{summary}{rforest}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{rforest}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the rforest function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/rforest.html} for an example in Radiant
}
\examples{
result <- rforest(titanic, "survived", "pclass", lev = "Yes")
summary(result)

}
\seealso{
\code{\link{rforest}} to generate results

\code{\link{plot.rforest}} to plot results

\code{\link{predict.rforest}} for prediction
}
