% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{avar}
\alias{avar}
\title{Analysis of Variance (ANOVA)}
\usage{
avar(
  dataset,
  treat,
  response,
  treats = "",
  ways = "one",
  model = "one",
  interaction = "FALSE",
  sum_check = "",
  data_filter = "",
  envir = parent.frame()
)
}
\arguments{
\item{dataset}{Dataset}

\item{treat}{The treatment variables}

\item{response}{The response variable}

\item{treats}{The treatment variables 2}

\item{model}{Type of ANOVA Model}

\item{interaction}{With Interaction}

\item{sum_check}{Optional output. "tukey" to show the Tukey Multiple Pairwise Comparisons. "posthoc" to show simultaneous tests for General Linear Hypotheses.}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}

\item{envir}{Environment to extract data from}

\item{way}{One Way or Two Way ANOVA}
}
\value{
A list of all variables used in the one-way anova function as an object of class avar
}
\description{
Analysis of Variance (ANOVA)
}
\details{
Anova function ...
}
\seealso{
\code{\link{summary.avar}} to summarize results

\code{\link{plot.avar}} to plot results
}
