% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{Catalog sales for men's and women's apparel}
\format{A data frame with 200 rows and 5 variables}
\usage{
data(catalog)
}
\description{
Catalog sales for men's and women's apparel
}
\details{
Description provided in attr(catalog, "description")
}
\keyword{datasets}
