% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalreg.R
\name{plot.evalreg}
\alias{plot.evalreg}
\title{Plot method for the evalreg function}
\usage{
\method{plot}{evalreg}(x, vars = c("Rsq", "RSME", "MAE"), ...)
}
\arguments{
\item{x}{Return value from \code{\link{evalreg}}}

\item{vars}{Measures to plot, i.e., one or more of "Rsq", "RSME", "MAE"}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the evalreg function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/evalreg.html} for an example in Radiant
}
\seealso{
\code{\link{evalreg}} to generate results

\code{\link{summary.evalreg}} to summarize results
}
