% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulater.R
\name{plot.repeater}
\alias{plot.repeater}
\title{Plot repeated simulation}
\usage{
\method{plot}{repeater}(x, shiny = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{repeater}}}

\item{shiny}{Did the function call originate inside a shiny app}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot repeated simulation
}

