% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.model}
\alias{radiant.model}
\alias{radiant.model-package}
\alias{radiant.model}
\title{radiant.model}
\usage{
radiant.model()
}
\description{
radiant.model

Launch radiant.model in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.model()
}
}
