% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.model_window}
\alias{radiant.model_window}
\title{Launch radiant.model in an Rstudio window}
\usage{
radiant.model_window()
}
\description{
Launch radiant.model in an Rstudio window
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.model_window()
}
}
