% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{store.crs}
\alias{store.crs}
\title{Deprecated: Store method for the crs function}
\usage{
\method{store}{crs}(dataset, object, name, ...)
}
\arguments{
\item{dataset}{Dataset}

\item{object}{Return value from \code{\link{crs}}}

\item{name}{Name to assign to the dataset}

\item{...}{further arguments passed to or from other methods}
}
\description{
Deprecated: Store method for the crs function
}
\details{
Return recommendations See \url{https://radiant-rstats.github.io/docs/model/crs.html} for an example in Radiant
}
