% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{store.nb.predict}
\alias{store.nb.predict}
\title{Store predicted values generated in the nb function}
\usage{
\method{store}{nb.predict}(dataset, object, name = "pred_nb", ...)
}
\arguments{
\item{dataset}{Dataset to add predictions two}

\item{object}{Return value from model function}

\item{name}{Variable name(s) assigned to predicted values. If empty, the levels of the response variable will be used}

\item{...}{Additional arguments}
}
\description{
Store predicted values generated in the nb function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/nb.html} for an example in Radiant
}
\examples{
result <- nb(titanic, rvar = "survived", evar = c("pclass", "sex", "age"))
pred <- predict(result, pred_data = titanic)
titanic <- store(titanic, pred, name = c("Yes", "No"))

}
